kakaopay/ceoplus/self_applications/mobile-test-automation/requirements.txt
Appium-Python-Client==3.1.0
selenium==4.15.0
pytest==7.4.3
allure-pytest==2.13.2
pytest-xdist==3.5.0